* 'listings-views' of https://github.com/erikcaineolson/HARE:
  Cleans up flash message notification: moves close ('x') to far-right, decreases the height, and removes ellipse corners.
  Moves tabs to 'Menu' dropdown on small screens and limits field width for 'New Listing' and 'View All Listings'
  Titlecase category names and listing titles
  Cleans up 'Create New Listing' view and hides tabs on smaller screens. TODO: fix menu display on small screens.
  Corrects broken style. TODO: Move as much custom stuff as possible from main into foundation_and_overrides as overrides of existing classes (instead of all new classes).
  Add image url attribute to Listing
  Add icon to 'un-watch' listing button
  Add Category filtered results to Listing views
  Corrects tabs not properly linking.
  Removes reference to default.js
  Lines up 'Watch' and 'Request Donation' buttons.
  Corrects passing the current user to the show view.
  Fixes 'hidden' filters on index.
  Corrects formatting on Listings index view.

+++ app/controllers/categories_controller.rb
@@ -1,16 +1,13 @@
 class CategoriesController < ApplicationController
-  def show
-    session.delete(:listings_index)
-    category_filters = params[:categories].split(",")
-    @categories = Category.all.to_a
-    @listings = Listing.joins(:categories).where("name IN (?)", category_filters).to_a
-    render 'listings/index'
-  end
-
   def index
-    session.delete(:listings_index)
     @categories = Category.all.to_a
-    @listings = Listing.all.to_a
+    unless params[:categories] and
+      params[:categories].values.length != @categories.length
+        redirect_to listings_path and return
+    end
+    @category_filters = params[:categories].values
+    @listings = Listing.joins(:categories).where("name IN (?)", params[:categories].values)
+    session.delete(:listings_index)
     render 'listings/index'
   end
 end
